Fix check style
